Finish functionality for post, user route

+++ src/controllers/post.js
@@ -1,9 +1,16 @@
 import Post from "../models/post.js";
-import { StatusCodes } from "http-status-codes";
-import { HTTPResponse, InternalServerErrorResponse, NotFoundResponse, generateErrorResponse } from "../utils/error.js";
+import { ReasonPhrases, StatusCodes } from "http-status-codes";
+import {
+  HTTPResponse,
+  InternalServerErrorResponse,
+  NotFoundResponse,
+  generateErrorResponse,
+  ForbiddenResponse,
+} from "../utils/error.js";
 import { OKHTTPResponse } from "../utils/response.js";
 import * as path from "path";
 import { fileURLToPath } from "url";
+import post from "../models/post.js";
 
 const fileName = path.basename(fileURLToPath(import.meta.url)).slice(0, -3);
 
@@ -48,7 +55,7 @@ export async function deleteAllPosts(req, res) {
 
 export async function getPost(req, res) {
   try {
-    const postId = req.params.id;
+    const postId = req.params.postId;
     const foundPost = await Post.getPost(postId);
     let response;
     !foundPost
@@ -63,11 +70,23 @@ export async function getPost(req, res) {
 
 export async function updatePost(req, res) {
   try {
-    const postId = req.params.id;
+    const postId = req.params.postId;
+    let response;
+    //Check if post exists
+    const foundPost = await Post.getPost(postId);
+    if (!foundPost) {
+      response = new NotFoundResponse(undefined, `Post ${postId} not found`);
+      return res.status(response.statusCode).json(response);
+    }
+    //Check if loggedInUser own the post
+    if (foundPost.userId != req.loggedInId) {
+      response = ForbiddenResponse.withMessage("User can not modify post belongs to a different user.");
+      return res.status(response.statusCode).json(response);
+    }
     const formData = { title: req.body.title, content: req.body.content };
     const result = await Post.updatePost(postId, formData);
-    const response = new OKHTTPResponse("Successfully updated a post", result);
-    res.status(response.statusCode).json(response);
+    response = new OKHTTPResponse("Successfully updated a post", result);
+    return res.status(response.statusCode).json(response);
   } catch (err) {
     const response = generateErrorResponse(fileName, err);
     res.status(response.statusCode).json(response);
@@ -76,7 +95,23 @@ export async function updatePost(req, res) {
 
 export async function deletePost(req, res) {
   try {
-    const postId = req.params.id;
+    const postId = req.params.postId;
+    //Check if loggedin user is author of post
+    const foundPost = await Post.getPost(postId);
+    if (!foundPost) {
+      const response = new NotFoundResponse(undefined, `Post ${postId} not found`);
+      return res.status(response.statusCode).json(response);
+    }
+    //Check if loggedInUser own the post
+    if (foundPost.userId != req.loggedInId) {
+      return res.status(StatusCodes.METHOD_NOT_ALLOWED).json({
+        statusCode: StatusCodes.METHOD_NOT_ALLOWED,
+        data: {
+          name: ReasonPhrases.METHOD_NOT_ALLOWED,
+          message: "User can not delete post belongs to a different user.",
+        },
+      });
+    }
     const result = await Post.deletePost(postId);
     const response = new OKHTTPResponse("Successfully deleted a post", result);
     res.status(response.statusCode).json(response);
